requirements.txt
playwright==1.40.0
PyPDF2==3.0.1
python-docx==0.8.11
openai==1.3.7
python-dotenv==1.0.0
pyyaml==6.0.1
beautifulsoup4==4.12.2
lxml==4.9.3
nltk==3.8.1
spacy==3.7.2
regex==2023.10.3
asyncio==3.4.3